% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertGradeCodeBlockAddin.R
\name{insertGradeCodeBlockAddin}
\alias{insertGradeCodeBlockAddin}
\title{Insert Rmarkdown code block for grading}
\usage{
insertGradeCodeBlockAddin()
}
\description{
Inserts a code block with knitr \verb{output.class with }knitr\verb{ chunk options to use HTML classes "}grade-src\code{", "}grade-out\code{", "}grade-msg\code{", "}grade-warn\code{", and "}grade-err`" for styling source code, output, messages,
warnings, and errors, respectively.
}
