% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertGradeCommentAddin.R
\name{insertGradeCommentAddin}
\alias{insertGradeCommentAddin}
\title{Insert Rmarkdown div for grading comment}
\usage{
insertGradeCommentAddin()
}
\description{
Uses the pandoc \code{:::} ... \code{:::} syntax for divs,
with a tag \code{{.comment}}
}
