% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupGradeCSS.R
\name{insertGradeCSSBlockAddin}
\alias{insertGradeCSSBlockAddin}
\title{Insert YAML block to include \code{grading.css}}
\usage{
insertGradeCSSBlockAddin()
}
\description{
Adds a format block in the document header YAML to include
\code{grading.css} in HTML output.
}
