% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdGrading.R
\docType{package}
\name{rmdGrading}
\alias{rmdGrading-package}
\alias{rmdGrading}
\title{rmdGrading Addin}
\description{
RStudio addin to insert grading comments and code-blocks into
RMarkdown documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonathan-g/rmdGrading}
  \item Report bugs at \url{https://github.com/jonathan-g/rmdGrading/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Gilligan \email{jonathan.gilligan@vanderbilt.edu} (\href{https://orcid.org/0000-0003-1375-6686}{ORCID})

}
\keyword{internal}
