% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupGradeCSS.R
\name{setupGradeCSSAddin}
\alias{setupGradeCSSAddin}
\title{Sets up CSS for grading comments}
\usage{
setupGradeCSSAddin()
}
\description{
Creates a file `css/grading.css' in the project root directory and
inserts YAML block at the cursor position.
}
