% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertGradeCodeBlockAddin.R,
%   R/insertGradeCommentAddin.R
\name{insertGradeCommentAddin}
\alias{insertGradeCommentAddin}
\title{Insert Rmarkdown code block for grading}
\usage{
insertGradeCommentAddin()

insertGradeCommentAddin()
}
\description{
Inserts a code block with knitr \verb{output.class with }knitr\verb{ chunk options to use HTML classes "}grade-src\code{", "}grade-out\code{", "}grade-msg\code{", "}grade-warn\code{", and "}grade-err`" for styling source code, output, messages,
warnings, and errors, respectively.

Uses the pandoc \code{:::} ... \code{:::} syntax for divs,
with a tag \code{{.comment}}
}
